+++ src/sections/item-group/components/ItemGroupEditModal.tsx
@@ -1,32 +1,20 @@
 import { createResource } from 'solid-js'
-import { type Accessor, createEffect, createSignal, Show } from 'solid-js'
+import { type Accessor, createSignal, Show } from 'solid-js'
 
-import { isOverflow } from '~/legacy/utils/macroOverflow'
-import {
-  currentDayDiet,
-  targetDay,
-} from '~/modules/diet/day-diet/application/dayDiet'
 import { type Item } from '~/modules/diet/item/domain/item'
-import { handleNewItemGroup } from '~/modules/diet/item-group/application/itemGroupEditUtils'
-import { askUnlinkRecipe } from '~/modules/diet/item-group/application/itemGroupModals'
+import {
+  handleItemApply,
+  handleItemDelete,
+  handleNewItemGroup,
+} from '~/modules/diet/item-group/application/itemGroupEditUtils'
 import { useItemGroupClipboardActions } from '~/modules/diet/item-group/application/useItemGroupClipboardActions'
+import { useUnlinkRecipeIfNotFound } from '~/modules/diet/item-group/application/useUnlinkRecipeIfNotFound'
 import {
   isRecipedItemGroup,
   type ItemGroup,
 } from '~/modules/diet/item-group/domain/itemGroup'
 import { isSimpleSingleGroup } from '~/modules/diet/item-group/domain/itemGroup'
-import {
-  removeItemFromGroup,
-  updateItemInGroup,
-} from '~/modules/diet/item-group/domain/itemGroupOperations'
-import { type MacroNutrients } from '~/modules/diet/macro-nutrients/domain/macroNutrients'
-import { getMacroTargetForDay } from '~/modules/diet/macro-target/application/macroTarget'
 import { fetchRecipeById } from '~/modules/diet/recipe/application/recipe'
-import {
-  isTemplateItemRecipe,
-  type TemplateItem,
-} from '~/modules/diet/template-item/domain/templateItem'
-import { showError } from '~/modules/toast/application/toastManager'
 import { Modal } from '~/sections/common/components/Modal'
 import { useConfirmModalContext } from '~/sections/common/context/ConfirmModalContext'
 import {
@@ -43,7 +31,6 @@ import {
   useItemGroupEditContext,
 } from '~/sections/item-group/context/ItemGroupEditContext'
 import { ExternalTemplateSearchModal } from '~/sections/search/components/ExternalTemplateSearchModal'
-import { stringToDate } from '~/shared/utils/date'
 
 type EditSelection = { item: Item } | null
 const [editSelection, setEditSelection] = createSignal<EditSelection>(null)
@@ -87,6 +74,18 @@ const InnerItemGroupEditModal = (props: ItemGroupEditModalProps) => {
 
   // Handler factories (curried)
   const handleNewItemGroupHandler = handleNewItemGroup({ group, setGroup })
+  const handleItemApplyHandler = handleItemApply({
+    group,
+    persistentGroup,
+    setGroup,
+    setEditSelection,
+    showConfirmModal,
+  })
+  const handleItemDeleteHandler = handleItemDelete({
+    group,
+    setGroup,
+    setEditSelection,
+  })
 
   const [recipeSignal, { mutate: setRecipeSignal }] = createResource(
     async () => {
@@ -102,112 +101,14 @@ const InnerItemGroupEditModal = (props: ItemGroupEditModalProps) => {
     },
   )
 
-  createEffect(() => {
-    const group_ = group()
-    const groupHasRecipe = isRecipedItemGroup(group_)
-    console.debug('Group changed:', group())
-
-    if (groupHasRecipe) {
-      setTimeout(() => {
-        if (recipeSignal.state === 'ready' && recipeSignal() === null) {
-          setTimeout(() => {
-            askUnlinkRecipe(
-              'A receita atrelada a esse grupo não foi encontrada. Deseja desvincular o grupo da receita?',
-              {
-                showConfirmModal,
-                group,
-                setGroup,
-              },
-            )
-          }, 0)
-        }
-      }, 200)
-    }
+  useUnlinkRecipeIfNotFound({
+    group,
+    recipeSignal,
+    showConfirmModal,
+    setGroup,
   })
 
   const canApply = group().name.length > 0 && editSelection() === null
-
-  const handleItemApply = (item: TemplateItem) => {
-    const group_ = group()
-    // TODO:   Allow user to edit recipe.
-    // TODO:   Allow user to edit recipe inside a group.
-    if (isTemplateItemRecipe(item)) {
-      showError(
-        'Ainda não é possível editar receitas! Funcionalidade em desenvolvimento',
-      )
-      return
-    }
-
-    const checkOverflow = (property: keyof MacroNutrients) => {
-      const macroOverflowOptions = (() => {
-        const currentItem = editSelection()?.item
-        if (currentItem === undefined) {
-          return { enable: false }
-        }
-        const originalItem = persistentGroup().items.find(
-          (i: Item) => i.id === currentItem.id,
-        )
-        if (originalItem === undefined) {
-          showError('Item original não encontrado', {
-            audience: 'system',
-          })
-          return { enable: false }
-        }
-        return { enable: true, originalItem }
-      })()
-
-      const currentDayDiet_ = currentDayDiet()
-      const macroTarget_ = getMacroTargetForDay(stringToDate(targetDay()))
-
-      return isOverflow(item, property, {
-        currentDayDiet: currentDayDiet_,
-        macroTarget: macroTarget_,
-        macroOverflowOptions,
-      })
-    }
-
-    const onConfirm = () => {
-      console.debug(
-        `[ExternalItemEditModal] onApply: setting itemId=${item.id} to item=`,
-        item,
-      )
-      const newGroup: ItemGroup = updateItemInGroup(group_, item.id, item)
-
-      console.debug('newGroup', newGroup)
-      setGroup(newGroup)
-      setEditSelection(null)
-    }
-
-    const isOverflowing =
-      checkOverflow('carbs') || checkOverflow('protein') || checkOverflow('fat')
-    if (isOverflowing) {
-      showConfirmModal({
-        title: 'Macronutrientes excedem metas diárias',
-        body: 'Os macronutrientes desse item excedem as metas diárias. Deseja continuar mesmo assim?',
-        actions: [
-          {
-            text: 'Cancelar',
-            onClick: () => undefined,
-          },
-          {
-            text: 'Continuar',
-            primary: true,
-            onClick: onConfirm,
-          },
-        ],
-      })
-    } else {
-      onConfirm()
-    }
-  }
-
-  const handleItemDelete = (itemId: TemplateItem['id']) => {
-    const newGroup: ItemGroup = removeItemFromGroup(group(), itemId)
-
-    console.debug('newGroup', newGroup)
-    setGroup(newGroup)
-    setEditSelection(null)
-  }
 
   return (
     <Show when={recipeSignal.state === 'ready'} keyed>
@@ -247,8 +148,8 @@ const InnerItemGroupEditModal = (props: ItemGroupEditModalProps) => {
                 if (!originalItem) return { enable: false }
                 return { enable: true, originalItem }
               }}
-              onApply={handleItemApply}
-              onDelete={handleItemDelete}
+              onApply={handleItemApplyHandler}
+              onDelete={handleItemDeleteHandler}
               onClose={() => setEditSelection(null)}
             />
           )}
